# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for tracing TON blockchain transactions and analyzing balance flows",
        "title": "TON Transaction Tracer API",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "tracer.zaruchevskiy.ru",
    "basePath": "/",
    "paths": {
        "/api/v1/dns/resolve": {
            "post": {
                "description": "Resolve a .ton or .t.me domain to wallet address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dns"
                ],
                "summary": "Resolve TON DNS domain",
                "parameters": [
                    {
                        "description": "Resolve request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ResolveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResolveResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/trace": {
            "post": {
                "description": "Trace a transaction chain and return balance changes for all involved accounts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Trace a TON transaction",
                "parameters": [
                    {
                        "description": "Trace request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TraceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TraceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/transactions/recent": {
            "post": {
                "description": "Get recent transactions for an address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Get recent transactions",
                "parameters": [
                    {
                        "description": "Recent transactions request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RecentTransactionsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.RecentTransactionsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Check if the service is healthy",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.HealthResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AccountInfo": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "EQACkbxaojFxCOihpgl-Xh5yJdnR1lgBN-QnRz9xrZVEpZdf"
                },
                "balance_change_ton": {
                    "type": "string",
                    "example": "-0.115324419"
                },
                "network_fees_ton": {
                    "type": "string",
                    "example": "0.005692355"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid address format"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "models.HealthResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "healthy"
                },
                "version": {
                    "type": "string",
                    "example": "1.0.0"
                }
            }
        },
        "models.RecentTransactionsRequest": {
            "type": "object",
            "required": [
                "address"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "example": "EQACkbxaojFxCOihpgl-Xh5yJdnR1lgBN-QnRz9xrZVEpZdf"
                },
                "count": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "models.RecentTransactionsResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "EQACkbxaojFxCOihpgl-Xh5yJdnR1lgBN-QnRz9xrZVEpZdf"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TransactionItem"
                    }
                }
            }
        },
        "models.ResolveRequest": {
            "type": "object",
            "required": [
                "domain"
            ],
            "properties": {
                "domain": {
                    "type": "string",
                    "example": "foundation.ton"
                }
            }
        },
        "models.ResolveResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "EQAU_6rW5f5P3eZCANkI2zCgIJl6PqDLuHVs2LmNaLfSXwIa"
                },
                "domain": {
                    "type": "string",
                    "example": "foundation.ton"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.TraceRequest": {
            "type": "object",
            "required": [
                "address"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "example": "EQACkbxaojFxCOihpgl-Xh5yJdnR1lgBN-QnRz9xrZVEpZdf"
                },
                "hash": {
                    "type": "string",
                    "example": "0a1073ddaae99eed5a6de55cbd60aee18b1fc3deeee59414024679bf00172229"
                },
                "lt": {
                    "type": "integer",
                    "example": 63633416000001
                },
                "scan_depth": {
                    "type": "integer",
                    "example": 1000
                }
            }
        },
        "models.TraceResponse": {
            "type": "object",
            "properties": {
                "accounts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AccountInfo"
                    }
                },
                "original_sender": {
                    "type": "string",
                    "example": "EQACkbxaojFxCOihpgl-Xh5yJdnR1lgBN-QnRz9xrZVEpZdf"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "total_transactions": {
                    "type": "integer",
                    "example": 9
                }
            }
        },
        "models.TransactionItem": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string",
                    "example": "0.05"
                },
                "hash": {
                    "type": "string",
                    "example": "0a1073ddaae99eed5a6de55cbd60aee18b1fc3deeee59414024679bf00172229"
                },
                "lt": {
                    "type": "integer",
                    "example": 63633416000001
                },
                "timestamp": {
                    "type": "integer",
                    "example": 1699876543
                }
            }
        }
    }
}